pyproject.toml
[project]
name = "repoclinic"
version = "0.1.0"
description = "Stateful CrewAI pipeline for deterministic repository analysis"
readme = "README.md"
authors = [
    { name = "Prayag Bhattarai", email = "bhattaraiprayag@hotmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "crewai>=1.9.3",
    "orjson>=3.11.7",
    "pathspec>=1.0.4",
    "pydantic>=2.11.10",
    "pyyaml>=6.0.3",
    "rich>=14.3.3",
    "tiktoken>=0.12.0",
    "typer>=0.23.1",
]

[project.scripts]
repoclinic = "repoclinic:main"

[build-system]
requires = ["uv_build>=0.10.4,<0.11.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
